refactor: 출석하지 않은 결석 판단 위치 변경
